(SIMP-275) Allow simp-rake-helpers when Puppet 4
Before this commit, the skeleton's Gemfile used special logic to prevent
bundler from including simp-rake-helpers if the environment variable
PUPPET_VERSION was any version of Puppet 4.  This was a workaround to an
overly restrictive .gemspec in simp-rake-helpers, which was corrected in
1.0.10 (SIMP-208).

This commit removes the Puppet 4 restriction from skeleton/Gemfile.

SIMP-275 #comment Fixed SIMP-285
SIMP-285 #close #comment Gemfile allows simp-rake-helpers in Puppet 4