Update patch set 1
Patch Set 1: Code-Review-2 Verified-1

I'm surprised it hasn't already been removed, but we shouldn't include rspec-puppet coverage in the module skeleton at all.  

rspec-puppet's implementation of coverage is problematic because it is "resource coverage."  Module spec tests should test for `compile.with_all_deps` and 100% of a module's *logic*―but the presence of a given resource in the catalog is irrelevant and pursuing its coverage for its own sake is a testing anti-pattern that results in overly brittle tests and technical disappointments.

Patch-set: 1
Label: Code-Review=-2
Label: Verified=-1